+++ Sources/StreamChat/APIClient/Endpoints/ChannelEndpoints_Tests.swift
@@ -447,6 +447,26 @@ final class ChannelEndpoints_Tests: XCTestCase {
         
         XCTAssertEqual(AnyEndpoint(expectedEndpoint), AnyEndpoint(endpoint))
     }
+    
+    func test_loadPinnedMessages_buildsCorrectly() {
+        let cid = ChannelId.unique
+        let query = PinnedMessagesQuery(
+            pageSize: .unique,
+            pagination: .aroundTimestamp(.unique)
+        )
+        
+        let expectedEndpoint = Endpoint<PinnedMessagesPayload>(
+            path: "channels/" + cid.apiPath + "/pinned_messages",
+            method: .get,
+            queryItems: nil,
+            requiresConnectionId: false,
+            body: ["payload": query]
+        )
+        
+        let endpoint: Endpoint<PinnedMessagesPayload> = .pinnedMessages(cid: cid, query: query)
+        
+        XCTAssertEqual(AnyEndpoint(expectedEndpoint), AnyEndpoint(endpoint))
+    }
 }
 
 extension ChannelEditDetailPayload {